Refactor Path and its codec
