+++ data-model/src/relative_encodings/entry_area.rs
@@ -2,7 +2,7 @@
 
 use ufotofu::{BulkConsumer, BulkProducer};
 use ufotofu_codec::{
-    Decodable, DecodableCanonic, DecodableSync, DecodeError, DecodingWentWrong, Encodable,
+    Decodable, DecodableCanonic, DecodableSync, DecodeError, Blame, Encodable,
     EncodableKnownSize, EncodableSync, RelativeDecodable, RelativeDecodableCanonic,
     RelativeDecodableSync, RelativeEncodable, RelativeEncodableKnownSize, RelativeEncodableSync,
 };
@@ -77,7 +77,7 @@ where
 }
 
 impl<const MCL: usize, const MCC: usize, const MPL: usize, N, S, PD>
-    RelativeDecodable<(N, Area<MCL, MCC, MPL, S>), DecodingWentWrong>
+    RelativeDecodable<(N, Area<MCL, MCC, MPL, S>), Blame>
     for Entry<MCL, MCC, MPL, N, S, PD>
 where
     N: NamespaceId + Decodable,
@@ -92,7 +92,7 @@ where
     async fn relative_decode<P>(
         producer: &mut P,
         r: &(N, Area<MCL, MCC, MPL, S>),
-    ) -> Result<Self, DecodeError<P::Final, P::Error, DecodingWentWrong>>
+    ) -> Result<Self, DecodeError<P::Final, P::Error, Blame>>
     where
         P: BulkProducer<Item = u8>,
         Self: Sized,
@@ -171,7 +171,7 @@ where
 }
 
 impl<const MCL: usize, const MCC: usize, const MPL: usize, N, S, PD>
-    RelativeDecodableCanonic<(N, Area<MCL, MCC, MPL, S>), DecodingWentWrong, DecodingWentWrong>
+    RelativeDecodableCanonic<(N, Area<MCL, MCC, MPL, S>), Blame, Blame>
     for Entry<MCL, MCC, MPL, N, S, PD>
 where
     N: NamespaceId + DecodableCanonic,
@@ -181,7 +181,7 @@ where
     async fn relative_decode_canonic<P>(
         producer: &mut P,
         r: &(N, Area<MCL, MCC, MPL, S>),
-    ) -> Result<Self, DecodeError<P::Final, P::Error, DecodingWentWrong>>
+    ) -> Result<Self, DecodeError<P::Final, P::Error, Blame>>
     where
         P: BulkProducer<Item = u8>,
         Self: Sized,
@@ -278,7 +278,7 @@ where
 }
 
 impl<const MCL: usize, const MCC: usize, const MPL: usize, N, S, PD>
-    RelativeDecodableSync<(N, Area<MCL, MCC, MPL, S>), DecodingWentWrong>
+    RelativeDecodableSync<(N, Area<MCL, MCC, MPL, S>), Blame>
     for Entry<MCL, MCC, MPL, N, S, PD>
 where
     N: NamespaceId + DecodableSync,
