+++ earthstar/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod cinn25519;
-pub mod identity_id;
-pub mod namespace_id;
+// pub mod cinn25519;
+// pub mod identity_id;
+// pub mod namespace_id;
