+++ fuzz/fuzz_targets/enc_path2_rel_path2.rs
@@ -1,8 +1,8 @@
-use ufotofu_codec::{fuzz_relative_all, DecodingWentWrong};
+use ufotofu_codec::{fuzz_relative_all, Blame};
 use willow_data_model::Path;
 
 const MCL: usize = 2;
 const MCC: usize = 3;
 const MPL: usize = 3;
 
-fuzz_relative_all!(Path<MCL, MCC, MPL>; Path<MCL, MCC, MPL>; DecodingWentWrong; DecodingWentWrong);
+fuzz_relative_all!(Path<MCL, MCC, MPL>; Path<MCL, MCC, MPL>; Blame; Blame);
