+++ fuzz/fuzz_targets/enc_path_rel_path.rs
@@ -1,8 +1,8 @@
-use ufotofu_codec::{fuzz_relative_all, DecodingWentWrong};
+use ufotofu_codec::{fuzz_relative_all, Blame};
 use willow_data_model::Path;
 
 const MCL: usize = 300;
 const MCC: usize = 300;
 const MPL: usize = 300;
 
-fuzz_relative_all!(Path<MCL, MCC, MPL>; Path<MCL, MCC, MPL>; DecodingWentWrong; DecodingWentWrong);
+fuzz_relative_all!(Path<MCL, MCC, MPL>; Path<MCL, MCC, MPL>; Blame; Blame);
