+++ fuzz/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod encode;
 pub mod path;
-pub mod placeholder_params;
+// pub mod placeholder_params;
 pub mod relative_codec_dummy;
-pub mod silly_sigs;
+// pub mod silly_sigs;
