+++ wgps/src/commitment_scheme/mod.rs
@@ -1,5 +1,5 @@
 use ufotofu::ConsumeAtLeastError;
-use willow_encoding::{Decodable, DecodeError, Encodable, RelationDecodable};
+// use willow_encoding::{Decodable, DecodeError, Encodable, RelationDecodable};
 
 use either::Either::*;
